move AllowReadingByAnyone into poolutil_test.go
